requirements.txt
pdfplumber
pymupdf
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
python-dateutil